Setup parser factory
