+++ .eslintignore
@@ -1,3 +1,2 @@
 /spec/fixtures/**
 /node_modules/**
-/lib/grammar.js
