+++ lib/assertions.js
@@ -1,6 +1,6 @@
 'use babel';
 
-import parse from './grammar';
+import parser from './grammar';
 import { matcherBuilder } from './matchers';
 
 
@@ -24,35 +24,10 @@ class Assertion {
 export class AssertionParser {
   constructor(iterator) {
     this.header = iterator.header;
-    this.regex = new RegExp(
-      `^(\\s*)${this.header.openToken}(.+?)${this.header.closeToken}$`
-    );
+    const { lex, parse } = parser(this.header.openToken, this.header.closeToken);
+    this.lex = lex;
+    this.parse = parse;
     this.iterator = iterator[Symbol.iterator]();
-  }
-
-  parseLine(line) {
-    const match = this.regex.exec(line);
-    if (!match) {
-      throw Error();
-    }
-
-    const [_match, leadingWhitespace, expression] = match;
-    const [positions, scopes] = parse.parse(expression);
-
-    return [
-      positions.map((p) => {
-        if (p <= 0) {
-          // start/end line indicators
-          return p;
-        } else if (p === 1) {
-          // open token indicator
-          return p + leadingWhitespace.length;
-        }
-        // carat indicator (includes open token, but move back 1 col to start of character)
-        return p + leadingWhitespace.length + this.header.openToken.length - 1;
-      }),
-      scopes,
-    ];
   }
 
   *[Symbol.iterator]() {
@@ -60,7 +35,7 @@ export class AssertionParser {
 
     for (const value of this.iterator) {
       try {
-        const [positions, [modifier, scopes]] = this.parseLine(value.line);
+        const [positions, [modifier, scopes]] = this.parse(this.lex(value.line));
         if (!currentLine) {
           throw Error("Can't have assertion before any syntax");
         }
