+++ lib/grammar.js
@@ -1,823 +1,243 @@
-module.exports = (function() {
-  "use strict";
+'use babel';
+/* eslint-disable new-cap */
 
-  /*
-   * Generated by PEG.js 0.9.0.
-   *
-   * http://pegjs.org/
-   */
+import { Lexer, Parser, Token } from 'chevrotain';
 
-  function peg$subclass(child, parent) {
-    function ctor() { this.constructor = child; }
-    ctor.prototype = parent.prototype;
-    child.prototype = new ctor();
+
+class Whitespace extends Token {
+  static PATTERN = /[ ]+/;
+}
+class Position extends Token {
+  static PATTERN = Lexer.NA;
+}
+class StartLinePosition extends Position {
+  static PATTERN = /<</;
+}
+class EndLinePosition extends Position {
+  static PATTERN = />>/;
+}
+class OpenTokenPosition extends Position {
+  static PATTERN = /<-/;
+}
+class CaratPosition extends Position {
+  static PATTERN = /\^/;
+}
+class Period extends Token {
+  static PATTERN = /[.]/;
+}
+class Identifier extends Token {
+  static PATTERN = /[a-zA-Z_][a-zA-Z0-9_-]*/;
+}
+class Modifier extends Token {
+  static PATTERN = Lexer.NA;
+}
+class Only extends Modifier {
+  static PATTERN = /only:|=/;
+}
+class OpenParens extends Token {
+  static PATTERN = /\(/;
+}
+class CloseParens extends Token {
+  static PATTERN = /\)/;
+}
+
+
+function isCaratLookahead() {
+  const result = (
+    this.LA(1) instanceof CaratPosition
+  ) || (
+    this.LA(1) instanceof Whitespace && this.LA(2) instanceof CaratPosition
+  );
+
+  return result;
+}
+
+
+export default function(openToken, _closeToken = null) {
+  class OpenToken extends Token {
+    static PATTERN = new RegExp(openToken);
   }
 
-  function peg$SyntaxError(message, expected, found, location) {
-    this.message  = message;
-    this.expected = expected;
-    this.found    = found;
-    this.location = location;
-    this.name     = "SyntaxError";
+  const allTokens = [
+    Whitespace,
+    OpenToken,
+    StartLinePosition,
+    EndLinePosition,
+    OpenTokenPosition,
+    CaratPosition,
+    Only,
+    Identifier,
+    Period,
+    OpenParens,
+    CloseParens,
+  ];
 
-    if (typeof Error.captureStackTrace === "function") {
-      Error.captureStackTrace(this, peg$SyntaxError);
-    }
-  }
+  const lexer = new Lexer(allTokens);
 
-  peg$subclass(peg$SyntaxError, Error);
+  class GrammarParser extends Parser {
+    constructor(input) {
+      var col = 1;  // eslint-disable-line no-var
+      super(input, allTokens, false /* isErrorRecoveryEnabled */);
+      const $ = this;
+      const openTokenLength = '// '.length;
 
-  function peg$parse(input) {
-    var options = arguments.length > 1 ? arguments[1] : {},
-        parser  = this,
+      this.root = $.RULE('assertion', () => {
+        $.OPTION(() => $.SUBRULE($.capturedSpaces));
+        $.SUBRULE($.openToken);
+        const positions = $.SUBRULE($.positions);
+        const scopes = $.SUBRULE($.scopes);
+        return [positions, scopes];
+      });
 
-        peg$FAILED = {},
+      this.capturedSpaces = $.RULE('capturedSpaces', () => {
+        const spaces = $.CONSUME(Whitespace);
+        col = col + spaces.image.length;
+        return spaces;
+      });
+      this.spaces = $.RULE('spaces', () => $.CONSUME(Whitespace));
 
-        peg$startRuleFunctions = { assertion: peg$parseassertion },
-        peg$startRuleFunction  = peg$parseassertion,
+      this.openToken = $.RULE('openToken', () => {
+        const token = $.CONSUME(OpenToken);
+        col = col + token.image.length;
+      });
 
-        peg$c0 = /^[ \t]/,
-        peg$c1 = { type: "class", value: "[ \\t]", description: "[ \\t]" },
-        peg$c2 = " ",
-        peg$c3 = { type: "literal", value: " ", description: "\" \"" },
-        peg$c4 = function(spaces) { return spaces.join(''); },
-        peg$c5 = "<<",
-        peg$c6 = { type: "literal", value: "<<", description: "\"<<\"" },
-        peg$c7 = function() { return [0]; },
-        peg$c8 = ">>",
-        peg$c9 = { type: "literal", value: ">>", description: "\">>\"" },
-        peg$c10 = function() { return [-1]; },
-        peg$c11 = "<-",
-        peg$c12 = { type: "literal", value: "<-", description: "\"<-\"" },
-        peg$c13 = function() { return [col]; },
-        peg$c14 = "^",
-        peg$c15 = { type: "literal", value: "^", description: "\"^\"" },
-        peg$c16 = function(spaces, carat) { col = col + spaces.length + carat.length; return col; },
-        peg$c17 = /^[a-zA-Z_]/,
-        peg$c18 = { type: "class", value: "[a-zA-Z_]", description: "[a-zA-Z_]" },
-        peg$c19 = /^[a-zA-Z0-9_\-]/,
-        peg$c20 = { type: "class", value: "[a-zA-Z0-9_-]", description: "[a-zA-Z0-9_-]" },
-        peg$c21 = function(firstChar, suffix) { return firstChar + suffix.join(''); },
-        peg$c22 = "only:",
-        peg$c23 = { type: "literal", value: "only:", description: "\"only:\"" },
-        peg$c24 = "=",
-        peg$c25 = { type: "literal", value: "=", description: "\"=\"" },
-        peg$c26 = function() { return "="; },
-        peg$c27 = function(only) { return only },
-        peg$c28 = ".",
-        peg$c29 = { type: "literal", value: ".", description: "\".\"" },
-        peg$c30 = function(dot, id) { return dot + id; },
-        peg$c31 = function(prefix, suffix) { return prefix + suffix.join(''); },
-        peg$c32 = function(scope) { return scope; },
-        peg$c33 = function(scope) { return [scope]; },
-        peg$c34 = function(scope, scopes) { return Array.prototype.concat.call([scope], scopes); },
-        peg$c35 = "(",
-        peg$c36 = { type: "literal", value: "(", description: "\"(\"" },
-        peg$c37 = ")",
-        peg$c38 = { type: "literal", value: ")", description: "\")\"" },
-        peg$c39 = function(scopes) { return scopes; },
-        peg$c40 = function(modifier, scopes) { return [modifier, scopes] },
-        peg$c41 = function(scopes) { return ["@", scopes]; },
+      this.startPosition = $.RULE('startPosition', () => {
+        $.CONSUME(StartLinePosition);
+        return [0];
+      });
 
-        peg$currPos          = 0,
-        peg$savedPos         = 0,
-        peg$posDetailsCache  = [{ line: 1, column: 1, seenCR: false }],
-        peg$maxFailPos       = 0,
-        peg$maxFailExpected  = [],
-        peg$silentFails      = 0,
+      this.endPosition = $.RULE('endPosition', () => {
+        $.CONSUME(EndLinePosition);
+        return [-1];
+      });
 
-        peg$result;
+      this.openPosition = $.RULE('openPosition', () => {
+        $.CONSUME(OpenTokenPosition);
+        return [col - openTokenLength];
+      });
 
-    if ("startRule" in options) {
-      if (!(options.startRule in peg$startRuleFunctions)) {
-        throw new Error("Can't start parsing from rule \"" + options.startRule + "\".");
-      }
+      this.caratPosition = $.RULE('caratPosition', () => {
+        $.OPTION(() => $.SUBRULE($.capturedSpaces));
+        const carat = $.CONSUME(CaratPosition);
+        col = col + carat.image.length;
+        return col - 1;
+      });
 
-      peg$startRuleFunction = peg$startRuleFunctions[options.startRule];
-    }
+      this.caratPositions = $.RULE('caratPositions', () => {
+        const positions = [];
+        $.AT_LEAST_ONE(isCaratLookahead, () => {
+          const pos = $.SUBRULE(this.caratPosition);
+          positions.push(pos);
+        }, 'caratPosition');
+        return positions;
+      });
 
-    function text() {
-      return input.substring(peg$savedPos, peg$currPos);
-    }
+      this.positions = $.RULE('positions', () => $.OR([
+        { ALT: () => $.SUBRULE($.startPosition) },
+        { ALT: () => $.SUBRULE($.endPosition) },
+        { ALT: () => $.SUBRULE($.openPosition) },
+        { ALT: () => $.SUBRULE($.caratPositions) },
+      ]));
 
-    function location() {
-      return peg$computeLocation(peg$savedPos, peg$currPos);
-    }
 
-    function expected(description) {
-      throw peg$buildException(
-        null,
-        [{ type: "other", description: description }],
-        input.substring(peg$savedPos, peg$currPos),
-        peg$computeLocation(peg$savedPos, peg$currPos)
-      );
-    }
+      this.only = $.RULE('only', () => {
+        $.CONSUME(Only);
+        return '=';
+      });
 
-    function error(message) {
-      throw peg$buildException(
-        message,
-        null,
-        input.substring(peg$savedPos, peg$currPos),
-        peg$computeLocation(peg$savedPos, peg$currPos)
-      );
-    }
+      this.modifier = $.RULE('modifier', () => $.SUBRULE($.only));
 
-    function peg$computePosDetails(pos) {
-      var details = peg$posDetailsCache[pos],
-          p, ch;
+      this.scopeSuffix = $.RULE('scopeSuffix', () => {
+        const period = $.CONSUME(Period);
+        const identifier = $.CONSUME(Identifier);
+        return period.image + identifier.image;
+      });
 
-      if (details) {
-        return details;
-      } else {
-        p = pos - 1;
-        while (!peg$posDetailsCache[p]) {
-          p--;
-        }
-
-        details = peg$posDetailsCache[p];
-        details = {
-          line:   details.line,
-          column: details.column,
-          seenCR: details.seenCR
-        };
-
-        while (p < pos) {
-          ch = input.charAt(p);
-          if (ch === "\n") {
-            if (!details.seenCR) { details.line++; }
-            details.column = 1;
-            details.seenCR = false;
-          } else if (ch === "\r" || ch === "\u2028" || ch === "\u2029") {
-            details.line++;
-            details.column = 1;
-            details.seenCR = true;
-          } else {
-            details.column++;
-            details.seenCR = false;
-          }
-
-          p++;
-        }
-
-        peg$posDetailsCache[pos] = details;
-        return details;
-      }
-    }
-
-    function peg$computeLocation(startPos, endPos) {
-      var startPosDetails = peg$computePosDetails(startPos),
-          endPosDetails   = peg$computePosDetails(endPos);
-
-      return {
-        start: {
-          offset: startPos,
-          line:   startPosDetails.line,
-          column: startPosDetails.column
-        },
-        end: {
-          offset: endPos,
-          line:   endPosDetails.line,
-          column: endPosDetails.column
-        }
-      };
-    }
-
-    function peg$fail(expected) {
-      if (peg$currPos < peg$maxFailPos) { return; }
-
-      if (peg$currPos > peg$maxFailPos) {
-        peg$maxFailPos = peg$currPos;
-        peg$maxFailExpected = [];
-      }
-
-      peg$maxFailExpected.push(expected);
-    }
-
-    function peg$buildException(message, expected, found, location) {
-      function cleanupExpected(expected) {
-        var i = 1;
-
-        expected.sort(function(a, b) {
-          if (a.description < b.description) {
-            return -1;
-          } else if (a.description > b.description) {
-            return 1;
-          } else {
-            return 0;
-          }
+      this.scope = $.RULE('scope', () => {
+        const identifier = $.CONSUME(Identifier);
+        let scope = identifier.image;
+        $.MANY(() => {
+          const suffix = $.SUBRULE($.scopeSuffix);
+          scope = scope + suffix;
         });
+        return scope;
+      });
 
-        while (i < expected.length) {
-          if (expected[i - 1] === expected[i]) {
-            expected.splice(i, 1);
-          } else {
-            i++;
-          }
-        }
-      }
+      this.scopeWithSpace = $.RULE('scopeWithSpace', () => {
+        $.OPTION(() => $.SUBRULE($.spaces));
+        return $.SUBRULE($.scope);
+      });
 
-      function buildMessage(expected, found) {
-        function stringEscape(s) {
-          function hex(ch) { return ch.charCodeAt(0).toString(16).toUpperCase(); }
+      this.singleScope = $.RULE('singleScope', () => {
+        const scope = $.SUBRULE($.scope);
+        return [scope];
+      });
 
-          return s
-            .replace(/\\/g,   '\\\\')
-            .replace(/"/g,    '\\"')
-            .replace(/\x08/g, '\\b')
-            .replace(/\t/g,   '\\t')
-            .replace(/\n/g,   '\\n')
-            .replace(/\f/g,   '\\f')
-            .replace(/\r/g,   '\\r')
-            .replace(/[\x00-\x07\x0B\x0E\x0F]/g, function(ch) { return '\\x0' + hex(ch); })
-            .replace(/[\x10-\x1F\x80-\xFF]/g,    function(ch) { return '\\x'  + hex(ch); })
-            .replace(/[\u0100-\u0FFF]/g,         function(ch) { return '\\u0' + hex(ch); })
-            .replace(/[\u1000-\uFFFF]/g,         function(ch) { return '\\u'  + hex(ch); });
-        }
+      this.multiScopes = $.RULE('multiScopes', () => {
+        const scopes = [$.SUBRULE($.scope)];
+        $.MANY(() => {
+          const scope = $.SUBRULE($.scopeWithSpace);
+          scopes.push(scope);
+        });
+        return scopes;
+      });
 
-        var expectedDescs = new Array(expected.length),
-            expectedDesc, foundDesc, i;
+      this.groupedScopes = $.RULE('groupedScopes', () => {
+        $.CONSUME(OpenParens);
+        const scopes = $.SUBRULE($.multiScopes);
+        $.CONSUME(CloseParens);
+        return scopes;
+      });
 
-        for (i = 0; i < expected.length; i++) {
-          expectedDescs[i] = expected[i].description;
-        }
+      this.modifiedScopes = $.RULE('modifiedScopes', () => {
+        const modifier = $.SUBRULE($.modifier);
+        const scopes = $.OR([
+          { ALT: () => $.SUBRULE($.groupedScopes) },
+          { ALT: () => $.SUBRULE($.singleScope) },
+        ]);
 
-        expectedDesc = expected.length > 1
-          ? expectedDescs.slice(0, -1).join(", ")
-              + " or "
-              + expectedDescs[expected.length - 1]
-          : expectedDescs[0];
+        return [modifier, scopes];
+      });
 
-        foundDesc = found ? "\"" + stringEscape(found) + "\"" : "end of input";
+      this.unmodifiedScopes = $.RULE('unmodifiedScopes', () => {
+        const scopes = $.OR([
+          { ALT: () => $.SUBRULE($.groupedScopes) },
+          { ALT: () => $.SUBRULE($.multiScopes) },
+        ]);
+        return ['@', scopes];
+      });
 
-        return "Expected " + expectedDesc + " but " + foundDesc + " found.";
-      }
+      this.scopes = $.RULE('scopes', () => {
+        $.SUBRULE($.spaces);
+        return $.OR([
+          { ALT: () => $.SUBRULE($.modifiedScopes) },
+          { ALT: () => $.SUBRULE($.unmodifiedScopes) },
+        ]);
+      });
 
-      if (expected !== null) {
-        cleanupExpected(expected);
-      }
-
-      return new peg$SyntaxError(
-        message !== null ? message : buildMessage(expected, found),
-        expected,
-        found,
-        location
-      );
+      Parser.performSelfAnalysis(this);
     }
+  }
 
-    function peg$parseassertion() {
-      var s0, s1, s2;
-
-      s0 = peg$currPos;
-      s1 = peg$parsepositions();
-      if (s1 !== peg$FAILED) {
-        s2 = peg$parsescopes();
-        if (s2 !== peg$FAILED) {
-          s1 = [s1, s2];
-          s0 = s1;
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
+  function lex(input) {
+    const result = lexer.tokenize(input);
+    if (result.errors.length) {
+      throw Error(result.errors.join('\n'));
     }
+    return result.tokens;
+  }
 
-    function peg$parse_() {
-      var s0, s1;
-
-      s0 = [];
-      if (peg$c0.test(input.charAt(peg$currPos))) {
-        s1 = input.charAt(peg$currPos);
-        peg$currPos++;
-      } else {
-        s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c1); }
-      }
-      while (s1 !== peg$FAILED) {
-        s0.push(s1);
-        if (peg$c0.test(input.charAt(peg$currPos))) {
-          s1 = input.charAt(peg$currPos);
-          peg$currPos++;
-        } else {
-          s1 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c1); }
-        }
-      }
-
-      return s0;
+  function parse(tokens) {
+    const parser = new GrammarParser(tokens, allTokens);
+    const result = parser.root();
+    if (parser.errors.length) {
+      throw Error(parser.errors.join('\n'));
     }
-
-    function peg$parsespaces() {
-      var s0, s1, s2;
-
-      s0 = peg$currPos;
-      s1 = [];
-      if (input.charCodeAt(peg$currPos) === 32) {
-        s2 = peg$c2;
-        peg$currPos++;
-      } else {
-        s2 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c3); }
-      }
-      while (s2 !== peg$FAILED) {
-        s1.push(s2);
-        if (input.charCodeAt(peg$currPos) === 32) {
-          s2 = peg$c2;
-          peg$currPos++;
-        } else {
-          s2 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c3); }
-        }
-      }
-      if (s1 !== peg$FAILED) {
-        peg$savedPos = s0;
-        s1 = peg$c4(s1);
-      }
-      s0 = s1;
-
-      return s0;
-    }
-
-    function peg$parsestartLinePosition() {
-      var s0, s1;
-
-      s0 = peg$currPos;
-      if (input.substr(peg$currPos, 2) === peg$c5) {
-        s1 = peg$c5;
-        peg$currPos += 2;
-      } else {
-        s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c6); }
-      }
-      if (s1 !== peg$FAILED) {
-        peg$savedPos = s0;
-        s1 = peg$c7();
-      }
-      s0 = s1;
-
-      return s0;
-    }
-
-    function peg$parseendLinePosition() {
-      var s0, s1;
-
-      s0 = peg$currPos;
-      if (input.substr(peg$currPos, 2) === peg$c8) {
-        s1 = peg$c8;
-        peg$currPos += 2;
-      } else {
-        s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c9); }
-      }
-      if (s1 !== peg$FAILED) {
-        peg$savedPos = s0;
-        s1 = peg$c10();
-      }
-      s0 = s1;
-
-      return s0;
-    }
-
-    function peg$parseopenTokenPosition() {
-      var s0, s1;
-
-      s0 = peg$currPos;
-      if (input.substr(peg$currPos, 2) === peg$c11) {
-        s1 = peg$c11;
-        peg$currPos += 2;
-      } else {
-        s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c12); }
-      }
-      if (s1 !== peg$FAILED) {
-        peg$savedPos = s0;
-        s1 = peg$c13();
-      }
-      s0 = s1;
-
-      return s0;
-    }
-
-    function peg$parsecaratPosition() {
-      var s0, s1, s2;
-
-      s0 = peg$currPos;
-      s1 = peg$parsespaces();
-      if (s1 !== peg$FAILED) {
-        if (input.charCodeAt(peg$currPos) === 94) {
-          s2 = peg$c14;
-          peg$currPos++;
-        } else {
-          s2 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c15); }
-        }
-        if (s2 !== peg$FAILED) {
-          peg$savedPos = s0;
-          s1 = peg$c16(s1, s2);
-          s0 = s1;
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
-    }
-
-    function peg$parsepositions() {
-      var s0, s1;
-
-      s0 = peg$parsestartLinePosition();
-      if (s0 === peg$FAILED) {
-        s0 = peg$parseendLinePosition();
-        if (s0 === peg$FAILED) {
-          s0 = peg$parseopenTokenPosition();
-          if (s0 === peg$FAILED) {
-            s0 = [];
-            s1 = peg$parsecaratPosition();
-            if (s1 !== peg$FAILED) {
-              while (s1 !== peg$FAILED) {
-                s0.push(s1);
-                s1 = peg$parsecaratPosition();
-              }
-            } else {
-              s0 = peg$FAILED;
-            }
-          }
-        }
-      }
-
-      return s0;
-    }
-
-    function peg$parseidentifier() {
-      var s0, s1, s2, s3;
-
-      s0 = peg$currPos;
-      if (peg$c17.test(input.charAt(peg$currPos))) {
-        s1 = input.charAt(peg$currPos);
-        peg$currPos++;
-      } else {
-        s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c18); }
-      }
-      if (s1 !== peg$FAILED) {
-        s2 = [];
-        if (peg$c19.test(input.charAt(peg$currPos))) {
-          s3 = input.charAt(peg$currPos);
-          peg$currPos++;
-        } else {
-          s3 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c20); }
-        }
-        while (s3 !== peg$FAILED) {
-          s2.push(s3);
-          if (peg$c19.test(input.charAt(peg$currPos))) {
-            s3 = input.charAt(peg$currPos);
-            peg$currPos++;
-          } else {
-            s3 = peg$FAILED;
-            if (peg$silentFails === 0) { peg$fail(peg$c20); }
-          }
-        }
-        if (s2 !== peg$FAILED) {
-          peg$savedPos = s0;
-          s1 = peg$c21(s1, s2);
-          s0 = s1;
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
-    }
-
-    function peg$parseonly() {
-      var s0, s1;
-
-      s0 = peg$currPos;
-      if (input.substr(peg$currPos, 5) === peg$c22) {
-        s1 = peg$c22;
-        peg$currPos += 5;
-      } else {
-        s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c23); }
-      }
-      if (s1 === peg$FAILED) {
-        if (input.charCodeAt(peg$currPos) === 61) {
-          s1 = peg$c24;
-          peg$currPos++;
-        } else {
-          s1 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c25); }
-        }
-      }
-      if (s1 !== peg$FAILED) {
-        peg$savedPos = s0;
-        s1 = peg$c26();
-      }
-      s0 = s1;
-
-      return s0;
-    }
-
-    function peg$parsemodifier() {
-      var s0, s1;
-
-      s0 = peg$currPos;
-      s1 = peg$parseonly();
-      if (s1 !== peg$FAILED) {
-        peg$savedPos = s0;
-        s1 = peg$c27(s1);
-      }
-      s0 = s1;
-
-      return s0;
-    }
-
-    function peg$parsescopeSuffix() {
-      var s0, s1, s2;
-
-      s0 = peg$currPos;
-      if (input.charCodeAt(peg$currPos) === 46) {
-        s1 = peg$c28;
-        peg$currPos++;
-      } else {
-        s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c29); }
-      }
-      if (s1 !== peg$FAILED) {
-        s2 = peg$parseidentifier();
-        if (s2 !== peg$FAILED) {
-          peg$savedPos = s0;
-          s1 = peg$c30(s1, s2);
-          s0 = s1;
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
-    }
-
-    function peg$parsescope() {
-      var s0, s1, s2, s3;
-
-      s0 = peg$currPos;
-      s1 = peg$parseidentifier();
-      if (s1 !== peg$FAILED) {
-        s2 = [];
-        s3 = peg$parsescopeSuffix();
-        while (s3 !== peg$FAILED) {
-          s2.push(s3);
-          s3 = peg$parsescopeSuffix();
-        }
-        if (s2 !== peg$FAILED) {
-          peg$savedPos = s0;
-          s1 = peg$c31(s1, s2);
-          s0 = s1;
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
-    }
-
-    function peg$parsescopeWithSpace() {
-      var s0, s1, s2;
-
-      s0 = peg$currPos;
-      s1 = peg$parse_();
-      if (s1 !== peg$FAILED) {
-        s2 = peg$parsescope();
-        if (s2 !== peg$FAILED) {
-          peg$savedPos = s0;
-          s1 = peg$c32(s2);
-          s0 = s1;
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
-    }
-
-    function peg$parsesingleScope() {
-      var s0, s1;
-
-      s0 = peg$currPos;
-      s1 = peg$parsescope();
-      if (s1 !== peg$FAILED) {
-        peg$savedPos = s0;
-        s1 = peg$c33(s1);
-      }
-      s0 = s1;
-
-      return s0;
-    }
-
-    function peg$parsemultiScopes() {
-      var s0, s1, s2, s3;
-
-      s0 = peg$currPos;
-      s1 = peg$parsescope();
-      if (s1 !== peg$FAILED) {
-        s2 = [];
-        s3 = peg$parsescopeWithSpace();
-        if (s3 !== peg$FAILED) {
-          while (s3 !== peg$FAILED) {
-            s2.push(s3);
-            s3 = peg$parsescopeWithSpace();
-          }
-        } else {
-          s2 = peg$FAILED;
-        }
-        if (s2 !== peg$FAILED) {
-          peg$savedPos = s0;
-          s1 = peg$c34(s1, s2);
-          s0 = s1;
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
-    }
-
-    function peg$parsegroupedScopes() {
-      var s0, s1, s2, s3;
-
-      s0 = peg$currPos;
-      if (input.charCodeAt(peg$currPos) === 40) {
-        s1 = peg$c35;
-        peg$currPos++;
-      } else {
-        s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c36); }
-      }
-      if (s1 !== peg$FAILED) {
-        s2 = peg$parsemultiScopes();
-        if (s2 === peg$FAILED) {
-          s2 = peg$parsesingleScope();
-        }
-        if (s2 !== peg$FAILED) {
-          if (input.charCodeAt(peg$currPos) === 41) {
-            s3 = peg$c37;
-            peg$currPos++;
-          } else {
-            s3 = peg$FAILED;
-            if (peg$silentFails === 0) { peg$fail(peg$c38); }
-          }
-          if (s3 !== peg$FAILED) {
-            peg$savedPos = s0;
-            s1 = peg$c39(s2);
-            s0 = s1;
-          } else {
-            peg$currPos = s0;
-            s0 = peg$FAILED;
-          }
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
-    }
-
-    function peg$parsemodifiedScopes() {
-      var s0, s1, s2, s3;
-
-      s0 = peg$currPos;
-      s1 = peg$parse_();
-      if (s1 !== peg$FAILED) {
-        s2 = peg$parsemodifier();
-        if (s2 !== peg$FAILED) {
-          s3 = peg$parsegroupedScopes();
-          if (s3 === peg$FAILED) {
-            s3 = peg$parsesingleScope();
-          }
-          if (s3 !== peg$FAILED) {
-            peg$savedPos = s0;
-            s1 = peg$c40(s2, s3);
-            s0 = s1;
-          } else {
-            peg$currPos = s0;
-            s0 = peg$FAILED;
-          }
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
-    }
-
-    function peg$parseunmodifiedScopes() {
-      var s0, s1, s2;
-
-      s0 = peg$currPos;
-      s1 = peg$parse_();
-      if (s1 !== peg$FAILED) {
-        s2 = peg$parsegroupedScopes();
-        if (s2 === peg$FAILED) {
-          s2 = peg$parsemultiScopes();
-          if (s2 === peg$FAILED) {
-            s2 = peg$parsesingleScope();
-          }
-        }
-        if (s2 !== peg$FAILED) {
-          peg$savedPos = s0;
-          s1 = peg$c41(s2);
-          s0 = s1;
-        } else {
-          peg$currPos = s0;
-          s0 = peg$FAILED;
-        }
-      } else {
-        peg$currPos = s0;
-        s0 = peg$FAILED;
-      }
-
-      return s0;
-    }
-
-    function peg$parsescopes() {
-      var s0;
-
-      s0 = peg$parsemodifiedScopes();
-      if (s0 === peg$FAILED) {
-        s0 = peg$parseunmodifiedScopes();
-      }
-
-      return s0;
-    }
-
-     
-      /**
-       * Column counter
-       * @type {Number}
-       */
-      var col = 1;
-
-
-    peg$result = peg$startRuleFunction();
-
-    if (peg$result !== peg$FAILED && peg$currPos === input.length) {
-      return peg$result;
-    } else {
-      if (peg$result !== peg$FAILED && peg$currPos < input.length) {
-        peg$fail({ type: "end", description: "end of input" });
-      }
-
-      throw peg$buildException(
-        null,
-        peg$maxFailExpected,
-        peg$maxFailPos < input.length ? input.charAt(peg$maxFailPos) : null,
-        peg$maxFailPos < input.length
-          ? peg$computeLocation(peg$maxFailPos, peg$maxFailPos + 1)
-          : peg$computeLocation(peg$maxFailPos, peg$maxFailPos)
-      );
-    }
+    return result;
   }
 
   return {
-    SyntaxError: peg$SyntaxError,
-    parse:       peg$parse
+    lex,
+    parse,
   };
-})();
+}
