+++ spec/grammar-spec.js
@@ -1,162 +1,188 @@
 'use babel';
 
-import parse from '../lib/grammar';
+import parser from '../lib/grammar';
 import { escapeSpecName } from '../lib/jasmine';
 
+const { lex, parse } = parser('// ');
+
+
+function itShouldLex(expression) {
+  it(`should lex: ${escapeSpecName(expression)}`, () => {
+    expect(() => lex(expression)).not.toThrow();
+  });
+}
+
+function itShouldNotLex(expression) {
+  it(`should not lex: ${escapeSpecName(expression)}`, () => {
+    expect(() => lex(expression)).toThrow();
+  });
+}
 
 function itShouldParse(expression) {
   it(`should parse: ${escapeSpecName(expression)}`, () => {
-    expect(() => parse.parse(expression)).not.toThrow();
+    expect(() => parse(lex(expression))).not.toThrow();
   });
 }
 
 function itShouldNotParse(expression) {
   it(`should not parse: ${escapeSpecName(expression)}`, () => {
-    expect(() => parse.parse(expression)).toThrow();
+    expect(() => parse(lex(expression))).toThrow();
   });
 }
 
 
 describe('Grammar', () => {
+  describe('Lexing', () => {
+    itShouldLex('// <- ');
+    itShouldLex('   // <- ');
+    itShouldLex('// << ');
+    itShouldLex('// >> ');
+    itShouldLex('// <- something');
+    itShouldLex('// <-        something');
+    itShouldLex('// <- =something');
+    itShouldLex('// <- only:something');
+    itShouldLex('// <- only:something.else');
+    itShouldLex('// <- only:something.else');
+    itShouldLex('// <- only:(something.else another)');
+    itShouldLex('// <- only:(something.else)');
+
+    itShouldNotLex('##');
+    itShouldNotLex('<>');
+    itShouldNotLex('//<- %something');
+    itShouldNotLex('// <- %something');
+    itShouldNotLex('// <- %something');
+  });
+
   describe('Parsing', () => {
-    itShouldParse('<- punctuation.definition.directive meta.preprocessor.c');
-    itShouldParse('    ^              ^   meta.brace.curly.js');
-    itShouldParse('    ^ blah.blah');
-    itShouldParse('    ^^^^    Something');
-    itShouldParse('<- =something');
-    itShouldParse('<- only:something');
-
-    itShouldNotParse('');
-    itShouldNotParse('nothing');
-    itShouldNotParse('^ ');
-    itShouldNotParse('<- ');
-    itShouldNotParse('<- =something else');
-    itShouldNotParse('<- only:something else');
-
+    describe('OpenToken', () => {
+      itShouldParse('// << something');
+      itShouldParse('   // << something');
+      itShouldParse('   // << something');
+    });
     describe('Positions', () => {
-      it('should parse start line operators', () => {
-        expect(parse.parse('<< something')).toEqual([
+      itShouldParse('// << something');
+      itShouldParse('// >> something');
+      itShouldParse('// <- something');
+      itShouldParse('//  ^ something');
+      itShouldParse('//    ^ something');
+      itShouldParse('// ^  ^ something');
+      itShouldParse('// ^^^^ something');
+      itShouldParse('// ^  ^ something.else another');
+
+      itShouldNotParse(' something');
+      itShouldNotParse('    something');
+      itShouldNotParse('//  something');
+      itShouldNotParse('   // something');
+      itShouldNotParse('// <> something');
+      it('should parse start of line', () => {
+        expect(parse(lex('  // << something'))).toEqual([
           [0],
           ['@', ['something']],
         ]);
       });
 
-      it('should parse end line operators', () => {
-        expect(parse.parse('>> something')).toEqual([
+      it('should parse end of line', () => {
+        expect(parse(lex('  // >> something'))).toEqual([
           [-1],
           ['@', ['something']],
         ]);
       });
 
-      it('should parse open token operators', () => {
-        expect(parse.parse('<- something')).toEqual([
-          [1],
+      it('should parse prefixed whitespace', () => {
+        expect(parse(lex('  // <- something'))).toEqual([
+          [3],
           ['@', ['something']],
         ]);
       });
 
-      it('should parse carat operators', () => {
-        expect(parse.parse('    ^ something')).toEqual([
+      it('should parse single carat', () => {
+        expect(parse(lex('  // ^ something'))).toEqual([
           [6],
           ['@', ['something']],
         ]);
       });
 
-      it('should parse multiple carat operators', () => {
-        expect(parse.parse(' ^  ^ something')).toEqual([
-          [3, 6],
-          ['@', ['something']],
-        ]);
-      });
-
-      it('should parse consecutive carat operators', () => {
-        expect(parse.parse(' ^^^^ something')).toEqual([
-          [3, 4, 5, 6],
+      it('should parse multiple carats', () => {
+        expect(parse(lex('  // ^   ^ something'))).toEqual([
+          [6, 10],
           ['@', ['something']],
         ]);
       });
     });
 
     describe('Scopes', () => {
-      it('should parse single part scope', () => {
-        expect(parse.parse('<- something')).toEqual([
-          [1],
+      itShouldParse('// << something');
+      itShouldParse('// << something.else');
+      itShouldParse('// << something another');
+      itShouldParse('// << something another.else');
+      itShouldParse('// << (something)');
+      itShouldParse('// << (something.else)');
+      itShouldParse('// << (something another.else)');
+      itShouldParse('// << =(something)');
+      itShouldParse('// << =(something.else)');
+      itShouldParse('// << =(something another.else)');
+      itShouldParse('// << only:(something)');
+      itShouldParse('// << only:(something.else)');
+      itShouldParse('// << only:(something another.else)');
+
+      itShouldNotParse('// << something.');
+      itShouldNotParse('// << .something');
+      itShouldNotParse('// << .');
+      itShouldNotParse('// << ()');
+      itShouldNotParse('// << =');
+      itShouldNotParse('// << =()');
+      itShouldNotParse('// << =:something another');
+      itShouldNotParse('// << only:');
+      itShouldNotParse('// << only:()');
+      itShouldNotParse('// << only:something another');
+
+      it('should parse a single scope', () => {
+        expect(parse(lex('// << something'))).toEqual([
+          [0],
           ['@', ['something']],
         ]);
       });
 
-      it('should parse single part scope with a modifier', () => {
-        expect(parse.parse('<- =something')).toEqual([
-          [1],
-          ['=', ['something']],
-        ]);
-      });
-
-      it('should parse a multipart scope', () => {
-        expect(parse.parse('<- something.else')).toEqual([
-          [1],
+      it('should parse a single nested scope', () => {
+        expect(parse(lex('// << something.else'))).toEqual([
+          [0],
           ['@', ['something.else']],
         ]);
       });
 
       it('should parse multiple scopes', () => {
-        expect(parse.parse('<- something else')).toEqual([
-          [1],
-          ['@', ['something', 'else']],
+        expect(parse(lex('// << something another'))).toEqual([
+          [0],
+          ['@', ['something', 'another']],
         ]);
       });
 
-
-      it('should parse grouped multiple scopes', () => {
-        expect(parse.parse('<- (something else.other)')).toEqual([
-          [1],
-          ['@', ['something', 'else.other']],
+      it('should parse a modified single scope', () => {
+        expect(parse(lex('// << =something'))).toEqual([
+          [0],
+          ['=', ['something']],
         ]);
       });
 
-      it('should parse a multipart scope with a dash', () => {
-        expect(parse.parse('<- something.my-attr else')).toEqual([
-          [1],
-          ['@', ['something.my-attr', 'else']],
+      it('should parse a modified grouped single scope', () => {
+        expect(parse(lex('// << only:(something)'))).toEqual([
+          [0],
+          ['=', ['something']],
         ]);
       });
 
-      describe('Modifiers', () => {
-        it('should parse single part scope', () => {
-          expect(parse.parse('<- =something')).toEqual([
-            [1],
-            ['=', ['something']],
-          ]);
-        });
+      it('should parse a single modified nested scope', () => {
+        expect(parse(lex('// << =something.else'))).toEqual([
+          [0],
+          ['=', ['something.else']],
+        ]);
+      });
 
-        it('should parse single part scope', () => {
-          expect(parse.parse('<- only:something')).toEqual([
-            [1],
-            ['=', ['something']],
-          ]);
-        });
-
-        it('should parse grouped scope', () => {
-          expect(parse.parse('<- =(something else)')).toEqual([
-            [1],
-            ['=', ['something', 'else']],
-          ]);
-        });
-
-        it('should parse grouped scope', () => {
-          expect(parse.parse('<- only:(something else)')).toEqual([
-            [1],
-            ['=', ['something', 'else']],
-          ]);
-        });
-
-        it('should not parse ungrouped scope', () => {
-          expect(() => parse.parse('<- =something else')).toThrow();
-        });
-
-        it('should not parse ungrouped scope', () => {
-          expect(() => parse.parse('<- only:something else')).toThrow();
-        });
+      it('should parse modified multiple scopes', () => {
+        expect(parse(lex('// << =(something another)'))).toEqual([
+          [0],
+          ['=', ['something', 'another']],
+        ]);
       });
     });
   });
